requirements.txt
numpy<=1.26.4
opencv-python<=4.6.0.66
Pillow<=10.4.0
PyYAML<=6.0.2
pyinstaller<6
picamera<=1.13
pre-commit<=3.8.0
GitPython<=3.1.43
